requirements copy.txt
paddle-rec==1.8.5.1